move dummy data structures to primitives
