+++ solo-chains/runtime/dancelight/src/bridge_to_ethereum_config.rs
@@ -23,16 +23,12 @@ use crate::symbiotic_message_processor::SymbioticMessageProcessor;
 #[cfg(not(test))]
 use crate::EthereumBeaconClient;
 use frame_support::weights::ConstantMultiplier;
-use parity_scale_codec::Encode;
-use snowbridge_router_primitives::inbound::{
-    ConvertMessage, ConvertMessageError, VersionedXcmMessage,
-};
+
+use sp_core::H160;
 use sp_core::{ConstU32, ConstU8};
-use sp_core::{H160, H256};
-use xcm::latest::{Assets, Location, SendError, SendResult, SendXcm, Xcm, XcmHash};
 use {
     crate::{
-        parameter_types, weights, xcm_config, xcm_config::UniversalLocation, AccountId,
+        parameter_types, weights, xcm_config, xcm_config::UniversalLocation,
         AggregateMessageOrigin, Balance, Balances, EthereumInboundQueue, EthereumOutboundQueue,
         EthereumSystem, FixedU128, GetAggregateMessageOrigin, Keccak256, MessageQueue, Runtime,
         RuntimeEvent, TransactionByteFee, TreasuryAccount, WeightToFee, UNITS,
@@ -41,6 +37,7 @@ use {
     pallet_xcm::EnsureXcm,
     snowbridge_beacon_primitives::{Fork, ForkVersions},
     snowbridge_core::{gwei, meth, AllowSiblingsOnly, PricingParameters, Rewards},
+    tp_bridge::{DoNothingConvertMessage, DoNothingRouter},
 };
 
 // Ethereum Bridge
@@ -229,36 +226,6 @@ mod test_helpers {
         fn verify(_: &Log, _: &Proof) -> Result<(), VerificationError> {
             Ok(())
         }
-    }
-}
-
-pub struct DoNothingRouter;
-impl SendXcm for DoNothingRouter {
-    type Ticket = Xcm<()>;
-
-    fn validate(
-        _dest: &mut Option<Location>,
-        xcm: &mut Option<Xcm<()>>,
-    ) -> SendResult<Self::Ticket> {
-        Ok((xcm.clone().unwrap(), Assets::new()))
-    }
-    fn deliver(xcm: Xcm<()>) -> Result<XcmHash, SendError> {
-        let hash = xcm.using_encoded(sp_io::hashing::blake2_256);
-        Ok(hash)
-    }
-}
-
-pub struct DoNothingConvertMessage;
-
-impl ConvertMessage for DoNothingConvertMessage {
-    type Balance = Balance;
-    type AccountId = AccountId;
-
-    fn convert(
-        _: H256,
-        _message: VersionedXcmMessage,
-    ) -> Result<(Xcm<()>, Self::Balance), ConvertMessageError> {
-        Err(ConvertMessageError::UnsupportedVersion)
     }
 }
